requirements.txt
amqp==5.0.2
asgiref==3.4.1
billiard==3.6.3.0
boto3==1.16.47
botocore==1.19.47
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
country-converter
cryptography==3.3.1
defusedxml==0.7.0rc1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.44.0
django-braces==1.14.0
django-celery-results
django-cors-headers==3.6.0
django-dbbackup==3.3.0
django-extensions==3.1.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-material-icons==0.1.2
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
jellyfish==0.6.1
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
numpy
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.6
packaging==20.8
pandas
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
stripe==2.55.1
swapper==1.1.2.post1
tablib==3.0.0
typing-extensions==3.7.4.3
twilio==6.63.0
uritemplate==3.0.1
urllib3==1.26.2
us
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
